+++ src/components/Accordion/AccordionControlled.tsx
@@ -5,23 +5,24 @@ import { classNames } from "utilities/props";
 import useElementId from "hooks/useElementId";
 import AccordionContext from "./Accordion.context";
 import * as T from "./Accordion.types";
+import useHandlerRef from "hooks/useHandlerRef";
 
 const AccordionControlled: React.FC<T.ControlledProps> = (props) => {
 	const { children, onToggle, active, iconPosition, iconSize, className, attributes } = props;
 	const rootClassNames = classNames(className);
 	const id = useElementId();
+	const onToggleRef = useHandlerRef(onToggle);
 
 	const value = React.useMemo(
 		() => ({
 			triggerId: `${id}-trigger`,
 			contentId: `${id}-content`,
 			active,
-			onToggle,
+			onToggle: onToggleRef.current,
 			iconPosition,
 			iconSize,
 		}),
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-		[active]
+		[active, iconPosition, iconSize, id, onToggleRef]
 	);
 
 	return (